requirements.txt
aiogram
asyncpg
google-cloud-translate
google-cloud-texttospeech
apscheduler
